moved timer class to its own file
