+++ Python/MpProcess.py
@@ -4,6 +4,7 @@ import mediapipe as mp
 from smbus import SMBus
 import time
 from AnimationMethods import AnimationMethods
+from timer import Timer
 
 addr = 0x8 # bus address
 bus = SMBus(7) # indicates /dev/ic2-1
@@ -27,35 +28,6 @@ def map_range(val, inMin, inMax, outMin, outMax):
 def clamp_number(num, a, b):
   return max(min(num, max(a, b)), min(a, b))
 
-class Timer:
-    
-    def __init__(self, interval):
-        self.interval = interval
-        self.start_time = None
-        self.pause_time = None
-        
-    def start(self):
-        
-        if self.pause_time is not None: #resume a paused timer
-            self.start_time += time.monotonic() - self.paused_time
-            self.paused_time = None
-        else: #start a new timer
-            self.start_time = time.monotonic()
-            
-    def is_done(self):
-        if self.start_time is None:
-            return False
-        elapsed_time = time.monotonic() - self.start_time
-        return elapsed_time >= self.interval
-    
-    def pause(self):
-        self.pause_time = time.monotonic()
-        
-    def cancel(self):
-        self.start_time = None
-        self.pause_time = None
-
-    
 
 class MpProcess:
     """
